% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify.R
\name{amplify}
\alias{amplify}
\title{Create or modify a column with nesting structure}
\usage{
amplify(.data, ...)
}
\arguments{
\item{.data}{An object with the data.}

\item{...}{Name-value pairs.}
}
\description{
Create or modify a column with nesting structure
}
