% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{candidate_summary}
\alias{candidate_summary}
\title{Summarize number of points within candidate buffers}
\usage{
candidate_summary(cands)
}
\arguments{
\item{cands}{\code{data.frame} of associations between points and nest
candidates returned by \code{\link{get_candidates}}}
}
\value{
Returns \code{tibble} counting number of points within each
candidate buffer.
}
\description{
\code{candidate_summary} counts the number of points grouped within each
candidate buffer and arranges them in descending order.
}
