% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\title{Link data frames as a chain}
\usage{
chain(..., .class = NULL)
}
\arguments{
\item{...}{A name-value pair where value is a data frame. The supplied
data frame should be linked by a key. The link is determined automatically by
the intersection of the column names but may be set manually by \code{link}.}
}
\description{
Link data frames as a chain
}
\examples{
vs_info <- data.frame(engine = c(0, 1), info = c("flat", "unknown"))
chain(original = mtcars,
      vs_total = group_by(original, vs) \%>\%
                  summarise(across(everything(), sum)),
      vs_dict = data.frame(vs = c(0, 1), engine = c("V-shaped", "Straight")),
      vs_info = link(vs_info, original, .by = c(vs = "engine")))



}
