% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_input}
\alias{check_input}
\title{Check input data format}
\usage{
check_input(gps_data)
}
\arguments{
\item{gps_data}{\code{data.frame} of GPS data}
}
\description{
\code{check_input} checks that the data input in \code{find_nests} is in
the correct format.
}
\details{
The function checks that the input data includes burst, date-time,
and lat/long coordinates.
}
