% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisit_funs.R
\name{choose_overlapping}
\alias{choose_overlapping}
\title{Handle overlapping attempts}
\usage{
choose_overlapping(attempts)
}
\arguments{
\item{attempts}{\code{data.frame} of revisitation patterns of candidate
nests}
}
\value{
Returns \code{data.frame} of revisitation patterns filtered to
only include non-temporally overlapping candidate nests
}
\description{
\code{choose_overlapping} selects top candidate nesting attempt among those
that are temporally overlapping.
}
\details{
Within the function \code{nest_finder}, \code{choose_overlapping}
is used when \code{discard_overlapping = TRUE}.

If the list of nest candidates includes temporally overlapping nesting
attempts, only the candidate with the most visits is kept and the others
get discarded. This is based on the rationale that an individual cannot
simultaneously nest at more than one location. The location that is
visited the most is assumed to be the most likely true nest.
}
