% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.ordinal.R
\name{detect.ordinal}
\alias{detect.ordinal}
\title{Detect potential ordinal variable.}
\usage{
detect.ordinal(data, nOrdinal = 10)
}
\arguments{
\item{data}{A data set.}

\item{nOrdinal}{The minimal of values allowed before being flags as ordinal. Default is 10.}
}
\value{
A list of detected ordinal variables and the density of each value
}
\description{
Detect potential ordinal variables and compute the dentisity of each level on a gaussian distribution which can be taken into account for sampling techniques like \code{raplce.ordinal()}.
}
\examples{
set.seed(73)
mydata <- data.frame(Likert = sample(1:7, replace = TRUE, size = 20),
                     Bernouilli = sample(0:1, replace = TRUE, size = 20))
detect.ordinal(mydata)
}
\references{
À ajouter.
}
\author{
André Achim (Matlab)
P.-O. Caron (R)
}
