% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{dist_mat}
\alias{dist_mat}
\title{Calculate distance matrix between points in the data}
\usage{
dist_mat(dat)
}
\arguments{
\item{dat}{\code{data.frame} of movement data for a single burst. Needs to
include burst, date, long, lat}
}
\value{
Returns \code{data.table} with distance matrix.
}
\description{
\code{dist_mat} calculates pairwise distances between all points in the
data.
}
\details{
Distances are calculated using the function
\code{\link[geosphere]{distGeo}}. Takes advantage of \code{data.table}
for a fast implementation. Adapted from a post on \href{
https://stackoverflow.com/questions/36817423/how-to-efficiently-calculate-
distance-between-pair-of-coordinates-using-data-tab}{StackOverflow}.
}
