% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_funs.R
\name{estimate_outcomes}
\alias{estimate_outcomes}
\title{Estimates nesting outcomes}
\usage{
estimate_outcomes(fixes, visits, mcmc_params = list(burn_in = 1000,
  n_chain = 2, thin = 5, n_adapt = 1000, n_iter = 10000))
}
\arguments{
\item{fixes}{A matrix of the number of GPS fixes on each day of the nesting attempt
as returned by \code{\link{format_visits}}.}

\item{visits}{A matrix of the number of nest visits on each day of the nesting attempt
as returned by \code{\link{format_visits}}.}
}
\description{
\code{estimate_outcomes} does blah blah blah...
}
\details{
Data can be passed from \code{\link{format_visits}}... blah blah blah
}
