% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_funs.R
\name{explore_nests}
\alias{explore_nests}
\title{Launch Shiny app for nesting data exploration}
\usage{
explore_nests(gps_data)
}
\arguments{
\item{gps_data}{\code{data.frame} of movement data. Needs to include burst,
date, long, lat}
}
\value{
nothing
}
\description{
\code{explore_nests} launches a Shiny app for dynamic exploration and
visualization of nests from movement data
}
\details{
The function takes as input \code{gps_data} (see
\code{\link{find_nests}} or vignette for details on data format) and
launches a Shiny app that allows the user to dynamically explore
repeatedly visited locations for one burst at a time while interactively
tuning input parameters (see \code{\link{find_nests}} for the complete
list). Under the hood, \code{explore_nests} runs \code{find_nests}, and
then displays the results on a map.
}
