% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fromJson,character,missing-method}
\alias{fromJson,character,missing-method}
\alias{fromJson-char-miss-method}
\title{Get Nested (char-miss)}
\usage{
\S4method{fromJson}{character,missing}(input, where = parent.frame(),
  strict = FALSE, ...)
}
\arguments{
\item{input}{\code{\link{character}}.}

\item{where}{\code{\link{missing}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}:
\code{FALSE}:}
}
\value{
See method
   \code{\link[nestr]{fromJson-char-char-method}}
}
\description{
See generic: \code{\link[nestr]{fromJson}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/nestr}
}
\seealso{
\code{
   \link[nestr]{fromJson}
}
}

