% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fromJson,json,missing-method}
\alias{fromJson,json,missing-method}
\alias{fromJson-json-miss-method}
\title{Get Nested (json-miss)}
\usage{
\S4method{fromJson}{json,missing}(input, where = parent.frame(),
  strict = FALSE, ...)
}
\arguments{
\item{input}{\code{\link{json}}.}

\item{where}{\code{\link{missing}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}:
\code{FALSE}:}
}
\value{
\code{\link{ANY}}. Option value or for non-existing option
   (i.e. wrong \code{input}): \code{NULL} if \code{strict = FALSE} and an error
   if \code{strict = TRUE}.
}
\description{
See generic: \code{\link[nestr]{fromJson}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/nestr}
}
\seealso{
\code{
   \link[nestr]{fromJson}
}
}

