% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{get_candidates}
\alias{get_candidates}
\title{Get top candidate nests from possible competitors}
\usage{
get_candidates(dm, buffer, min_pts = 2)
}
\arguments{
\item{dm}{Distance matrix returned by \code{\link{dist_mat}}}

\item{buffer}{Buffer distance (in meters) used to group points}

\item{min_pts}{Minimum number of points within the buffer for a point to be
retained. Defaults to 2}
}
\value{
Returns \code{data.frame} relating original location identifiers
(\code{loc_id}) to the identifier of the corresponding candidate nest
(\code{group_id}).
}
\description{
\code{get_candidates} uses a distance matrix returned by
\code{\link{dist_mat}} and a user-defined buffer to select candidate nest
sites.
}
\details{
Due to both movement and GPS error, recorded points around
recurrently visited locations are expected to be scattered around the true
revisited location. The buffer is meant to account for this scattering,
by grouping points that fall within the buffer distance.

When grouping, several points peripheral to the true revisited location
may compete in grouping points around them. We term these 'competing
points'. If the buffers of two points do not overlap, those points are
not competing. Based on the assumption that the true revisited location
is the one that incorporates the most points within its buffer,
\code{get_candidates} compares the number of points that fall within the
buffers of competing points and selects the one that includes the most.

A top candidate is selected for each cluster of competing points, i.e., one
representative for each cluster around a revisited location. If there are
multiple revisited locations with non-competing points, independent top
candidates are all returned.

To speed up calculations, the user can define \code{min_pts} as the minimum
number of points that need to fall within the buffer for a point to be
considered as a potential candidate. This discards isolated points from
consideration as revisited locations.
}
