% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{get_candidates_multi}
\alias{get_candidates_multi}
\title{Get top candidate nests from possible competitors (multiple buffers)}
\usage{
get_candidates_multi(dmat, buffers, min_pts = 2)
}
\arguments{
\item{dmat}{Distance matrix returned by \code{\link{dist_mat}}}

\item{buffers}{Vector of buffer size (in meters) used to group points}

\item{min_pts}{Minimum number of points within the buffer for a point to be
retained. Defaults to 2}
}
\value{
Returns a \code{list} of data frames relating original location identifiers
(\code{loc_id}) to the identifier of the corresponding candidate nest
(\code{group_id}) for each of the specified buffers.
}
\description{
\code{get_candidates_multi} is an adaptation of \code{\link{get_candidates}}
for multiple buffers. It is called by \code{\link{compare_buffers}}.
}
