% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explo_funs.R
\name{get_explodata}
\alias{get_explodata}
\title{Extract nests and non-nests from revisited locations}
\usage{
get_explodata(candidate_nests, known_coords, known_ids, buffer,
  pick_overlapping = TRUE)
}
\arguments{
\item{candidate_nests}{\code{data.frame} of candidate nests output by
\code{find_nests}}

\item{known_coords}{\code{data.frame} of coordinates for known nests. Needs
to include burst, long, lat.}

\item{known_ids}{\code{data.frame} of location IDs for known nests. Needs
to include burst and loc_id.}

\item{buffer}{Integer. Buffer distance (in meters) used to select true
nest location among candidates when \code{known_coords} is provided.}

\item{pick_overlapping}{Logical. If \code{TRUE} (default), the non-nest is
picked among those whose time range overlaps with the true nesting attempt.}
}
\value{
A \code{data.frame} including an equal number of true nests and
non-nests and their revisitation parameters.
}
\description{
\code{get_explodata} uses known nest locations to extract an equal number of
true nests and non-nests from a set of revisited locations for exploration
of parameter values.
}
\details{
If no prior information is available to the user about which
parameter values to use to filter nests among revisited locations, some
exploration of the data is necessary. The objective of the exploration
phase is to identify the set of parameter values that best discriminates
between nests and non-nests in the species or population at hand.

Our suggested procedure consists in running a first coarse screening of
revisited locations, using loose thresholds of behavioral parameters for
filtering. For example, using \code{min_consec = 1}, \code{min_top_att = 1},
and \code{min_days_att = 1} in \code{find_nests}. In most cases, previous
knowledge about the biology of the species should allow the user to
already provide values for \code{sea_start}, \code{sea_end}, and
\code{nest_cycle}. The output of this first screening is a set of revisited
locations at the \code{buffer} distance of choice, which should include
some nests as well as other repeatedly visited locations that are not nests.
Comparing the values of behavioral parameters at nests versus non-nests can
inform the choice of parameter values for later analysis.

This procedure requires knowledge of true nest locations for at least a
subset of the data. Given some data on known nest locations and the
coarse-screening output of \code{find_nests}, the function
\code{get_explodata} identifies true nests and an equal number of non-nests
to compare them to.

Comparing behavioral parameter values at nests versus non-nests will
allow the user to find the set of parameter values that best discriminates
between them. The resulting set of parameters can then be used to find nests
in new data or in a subset of data for which no prior information on nests
is available.

The user can pass data on known nests as either coordinates or location IDs.
Ideally, prior and independent information on nest locations
is available for a subset of the data. In this case, we recommend
that coordinates are passed to the function argument \code{known_coords}.
When coordinates of true nests are not known a-priori but the user is able
to visually inspect revisited locations and identify those that are true
nests (for example because they fall within known colonies), providing
location IDs for true nests in \code{known_ids} is an alternative option.

When passing \code{known_coords}, the user is required to also specify a
value for \code{buffer}. Because of GPS error, the coordinates of the point
representing the true nest in \code{candidate_nests} might not exactly match
those of the known nest location. If coordinates of true nests are provided
rather than location IDs, the function selects the true nest among the set
of candidates by choosing the candidate with the most visits among those
that fall within a \code{buffer} distance from the known nest location.
We recommend using for this argument the same value used for argument
\code{buffer} in \code{find_nests}.
}
