% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funs.R
\name{julian_to_date}
\alias{julian_to_date}
\title{Conversion of Julian day to date}
\usage{
julian_to_date(jd, yr)
}
\arguments{
\item{jd}{Integer. Julian day to convert.}

\item{yr}{Integer. Year the Julian date refers to.}
}
\value{
Returns a date in ymd format.
}
\description{
\code{julian_to_date} returns a date corresponding to Julian day and year
in input.
}
\examples{
julian_to_date(1, 2000)
# Returns January 1st of 2000

# Test what happens with leap years

julian_to_date(100, 1991)
# 1991 was not a leap year; Julian day 100 is April 10th

julian_to_date(100, 1992)
# 1992 was a leap year; Julian day 100 is April 9th
}
