% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin_with}
\alias{margin_with}
\title{Calculate margin statistics}
\usage{
margin_with(.x, .f, ..., .group_by = NULL, .keyname = NULL)
}
\arguments{
\item{f}{A function to evaluate for the given group.}

\item{group_by}{The indexing group for \code{f}.}
}
\value{
A named list where name corresponds to group levels and
entries to the results of \code{f}.
}
\description{
Calculate margin statistics
}
\examples{
with(PlantGrowth, margin_with(weight, sum, .group_by = group))
}
