% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestR.R
\docType{package}
\name{nestR}
\alias{nestR}
\title{nestR: A package Analysis of movement recursions to detect reproductive events
and estimate their fate in central place foragers.}
\description{
Functions in nestR implement an approach first described in our manuscript,
"Analysis of movement recursions to detect reproductive events and estimate
their fate in central place foragers" (Picardi et al. 2020).
}
\details{
Find it here:
https://doi.org/10.1186/s40462-020-00201-1
}
