% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.unique.variable.R
\name{p.unique.variable}
\alias{p.unique.variable}
\title{Probability of unique variables}
\usage{
p.unique.variable(x)
}
\arguments{
\item{x}{A data frame with only numeric variables.}
}
\value{
A data frame containing the F-values and probabilities of the variable to be an unique variable.
}
\description{
Probability of unique variables
}
\examples{
exData <- genr8(n = 420, R = ex_3factors_doub_unique)
p.unique.variable(exData)
}
\author{
P.-O. Caron (R)
André Achim (Matlab)
}
