% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_funs.R
\name{perf_metrics}
\alias{perf_metrics}
\title{Compute nest-detection performance metrics}
\usage{
perf_metrics(gps_data, nest_info, known_coords, min_consec)
}
\arguments{
\item{gps_data}{Original \code{data.frame} of movement data. Needs to include
burst, date, long, lat}

\item{nest_info}{Output of \code{find_nests}}

\item{known_coords}{\code{data.frame} of coordinates for known nests. Needs
to include burst, long, lat}

\item{min_consec}{The minimum number of consecutive days visited that was
used when running \code{find_nests}}
}
\description{
\code{perf_metrics} calculates nest detection performance metrics
based on the results of \code{\link{find_nests}} and data on known
nest locations. The performance metrics are defined as follows:
}
\details{
\itemize{

\item Positive predictive value, i.e., the proportion of identified nests
that are known to be true nests;
\item Sensitivity, i.e., the proportion of known nests that were successfully
identified;
\item False negative rate, i.e., the proportion of known nests that we failed
to identify.
}

The false positive rate, i.e., the proportion of spurious nests among
those identified, is not computed because it would require the assumption
that any nests that are detected but are not known are spurious, which
is not reasonable in most situations.
}
