% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_funs.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival over time from MCMC run}
\usage{
plot_survival(mcmc_obj, ci = 0.95)
}
\arguments{
\item{mcmc_obj}{List of \code{mcarrays} output by \code{estimate_outcomes}.}

\item{ci}{Numeric. Credible interval level.}
}
\value{
A plot of nest survival over time.
}
\description{
\code{plot_survival} makes a plot of mean nest survival over time
as estimated by \code{estimate_outcomes}.
}
