% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.model.R
\name{prep.model}
\alias{prep.model}
\title{Prepare surrogate model for NEST}
\usage{
prep.model(R, factors)
}
\arguments{
\item{R}{A correlation matrix.}

\item{factors}{The number of factors.}
}
\value{
The
}
\description{
Prepare surrogate model for NEST
}
\examples{
prep.model(R = ex_4factors_corr, factors = 2)
}
