% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.puniquevar.R
\name{print.puniquevar}
\alias{print.puniquevar}
\title{Print output of \code{p.unique.variable()}}
\usage{
\method{print}{puniquevar}(x, ...)
}
\arguments{
\item{x}{An object of class "p.uniquevar".}

\item{...}{Further arguments for other methods, ignored for class "p.uniquevar".}
}
\description{
Print the results of \code{p.unique.variable()}
}
\examples{
exData <- genr8(n = 420, R = ex_3factors_doub_unique)
out <- p.unique.variable(exData)
print(out)
}
