% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.resym.R
\name{print.resym}
\alias{print.resym}
\title{Print output of \code{reSym()}}
\usage{
\method{print}{resym}(x, ...)
}
\arguments{
\item{x}{An object of class "resymr".}

\item{...}{Further arguments for other methods, ignored for classe "resym".}
}
\description{
Print the results of \code{p.reSym()}.
}
\examples{
n <- 200
A <- rnorm(n)
B <- rlnorm(n, sd = 2)
C <- -rlnorm(n, sd = 2)
x <- cbind(A,B,C)
round(apply(x, 2, skew), 3)
out <- reSym(x)
print(out)
}
