% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcfa.R
\name{rcfa}
\alias{rcfa}
\title{Regularized Common Factor Analysis}
\usage{
rcfa(R, factors, Lstart = 0.9)
}
\arguments{
\item{R}{The correlation matrix.}

\item{factors}{The number of factors.}

\item{Lstart}{An initial value.}
}
\value{
A list of 
\itemize{
  \item \code{loadings} -A matrix of unrotated factor loadings.
  \item \code{Rr} - The reduced corrlation matrix.
  \item \code{Vb} - A vector of estimated communality values. 
  \item \code{L} - Value of the estimated penality parameter.
}
}
\description{
Regularized Common Factor Analysis
}
\examples{
rcfa(R = ex_4factors_corr, factors = 4)
}
\references{
Jung, S. & Takane, Y. (2008). Regularized common factor analysis. \emph{New trends in psychometrics}, 141-149.
}
\author{
P.-O. Caron (R)
André Achim (Matlab)
}
