% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reactiveExpression}
\alias{reactiveExpression}
\title{Create Reactive Expression}
\usage{
reactiveExpression(expr, where = parent.frame(), ...)
}
\arguments{
\item{expr}{\code{\link{expression}} (quoted or unquoted).}

\item{where}{\code{\link{environment}}.
The parent environment for the reactive expression.
By default, this is the calling environment, the same as when defining
an ordinary non-reactive expression.}

\item{Further}{arguments to be passed along to subsequent functions.
In particular:
\code{\link[reactr]{reactiveExpression}}.}
}
\description{
Creates a reactive option expression. This is a convenience/encapsulation
wrapper around \code{\link[reactr]{reactiveExpression}}.
}
\note{
This is just a temporary solution for facilitating debugging the interaction
with and the use of \code{\link[reactr]{reactiveExpression}} whose use is
the ultimate goal. That is also why the import of \code{\link{shiny}} is
not clearly stated in \code{DESCRIPTION} and \code{NAMESPACE}.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/nestr}
}
\seealso{
\code{
    \link[reactr]{reactiveExpression}
}
}

