% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.ordinal.R
\name{replace.ordinal}
\alias{replace.ordinal}
\title{Convert continuous data into ordinal data}
\usage{
replace.ordinal(ord.var, D)
}
\arguments{
\item{ord.var}{A data frame with ordinal variables' name as column names and the probability of each ordinal values in the row.}

\item{D}{A data frame to be converted (normal to ordinal).}
}
\value{
A data set the same size as D.
}
\description{
Using the density of each level, convert a continuous variables into an ordinal variable. Useful with \code{detect.ordinal()}.
}
\examples{
set.seed(73)
mydata <- data.frame(Likert = sample(1:7, replace = TRUE, size = 20),
                     Bernouilli = sample(0:1, replace = TRUE, size = 20))
ord.var <- detect.ordinal(mydata)
D <- genr8(n = 20, R = diag(ncol(mydata)))
colnames(D) <- c("Likert", "Bernouilli") 
replace.ordinal(ord.var, D)
}
\references{
À ajouter.
}
\author{
André Achim (Matlab)
P.-O. Caron (R)
}
