% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisit_funs.R
\name{revisit_stats}
\alias{revisit_stats}
\title{Calculate revisitation patterns}
\usage{
revisit_stats(sub, sea_start, sea_end, min_d_fix, min_consec, nest_cycle)
}
\arguments{
\item{sub}{\code{data.frame}. Subset of movement data corresponding to
candidate nests}

\item{sea_start}{Integer (if Julian day) or date in which the nesting season
starts}

\item{sea_end}{Integer (if Julian day) or date in which the nesting season
ends}

\item{min_d_fix}{Integer. Minimum number of fixes in a day for that day to
be counted as not visited if a visit was not observed}
}
\value{
Returns \code{data.frame} with revisitation statistics for each
candidate nest.
}
\description{
\code{revisit_stats} calculates patterns of revisitation at candidate nests.
}
\details{
This is a wrapper function that calls \code{visit_rle},
\code{rle_to_consec}, and \code{attempt_limits}.

For each candidate nest, the function computes the first and last day when
the location was visited, the total number of visits, the number of days in
which it was visited, the percent of days with a visit, the attendance on
the day with the most visits (percent locations at the nest over the total
number of fixes on that day), the longest series of consecutive days
visited, and the estimated start and end dates of the nesting attempt.

On days when no visit was recorded, two cases are possible: either the nest
was truly not visited, or visits were missed. On days with few fixes, there
is a higher chance of missing a visit given that it happened. Missed visit
detections can interrupt an otherwise continuos strike of days visited.
To counteract possible issues due to missed visit detections, the user can
define \code{min_d_fix} to set a minimum number of fixes that have to be
available in a day with no visits for that day to be retained when counting
consecutive days visited. If a day with no visits and fewer fixes than
\code{min_d_fix} interrupts a sequence of consecutive days visited, it
does not get considered and the sequence gets counted as uninterrupted.
}
