% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisit_funs.R
\name{rle_to_consec}
\alias{rle_to_consec}
\title{Calculate consecutive days visited}
\usage{
rle_to_consec(rl_df)
}
\arguments{
\item{rl_df}{\code{data.frame} of run-length encoding of nest visits}
}
\value{
Returns maximum number of consecutive days when the candidate
nest was visited.
}
\description{
\code{rle_to_consec} calculates the longest sequence of consecutive days a
candidate nest was visited.
}
\details{
Used with \code{lapply} inside function \code{revisit_stats}.
Computes longest series of consecutive days a candidate nest was visited.
Takes as input the output of \code{visit_rle}.
}
