% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Calculate univariate skew of a variaable}
\usage{
skew(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The skewness of the variable.
}
\description{
Calculate univariate skew of a variaable
}
\examples{
set.seed(42)
x <- rnorm(420)^2
skew(x)
}
