% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_funs.R
\name{summarize_outcomes}
\alias{summarize_outcomes}
\title{Summary of outcomes from MCMC run}
\usage{
summarize_outcomes(mcmc_obj, ci = 0.95)
}
\arguments{
\item{mcmc_obj}{List of \code{mcarrays} output by \code{estimate_outcomes}.}

\item{ci}{Numeric. Credible interval level.}
}
\value{
A list with the population level survival and detection parameters and the
individual nest fates
}
\description{
\code{summarize_outcomes} blah blah
}
