% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisit_funs.R
\name{visit_rle}
\alias{visit_rle}
\title{Run-length encoding of nest visits}
\usage{
visit_rle(x)
}
\arguments{
\item{x}{\code{data.frame} of daily revisitation history of a
candidate nest}
}
\value{
Returns \code{data.frame} of run-length encoding of nest visits.
}
\description{
\code{visit_rle} calculates the run-length encoding of visits based on
the daily history of revisitation of a candidate nest.
}
\details{
Used with \code{lapply} inside function \code{revisit_stats}.
Performs run-length encoding of nest visits and formats as
\code{data.frame} for later use.
}
