% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataformat_funs.R
\name{format_attempts}
\alias{format_attempts}
\title{Format data for nesting outcome estimation}
\usage{
format_attempts(nest_info, nest_cycle)
}
\arguments{
\item{nest_info}{Output of \code{find_nests}}

\item{nest_cycle}{Duration of nesting cycle}
}
\description{
\code{format_attempts} takes as input the output of \code{find_nests} and
formats it for input in \code{estimate_outcomes}.
}
\details{
bla
}
