% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestr-package.R
\docType{package}
\name{nestr-package}
\alias{nestr}
\alias{nestr-package}
\title{nestr: Easily make data frame with nesting or hierarchical structures}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
