% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestr-package.R
\docType{package}
\name{nestr-package}
\alias{nestr}
\alias{nestr-package}
\title{nestr: Build Nesting or Hierarchical Structures}
\description{
Facilitates building a nesting or hierarchical structure as a list or data frame by using a human friendly syntax.
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
\keyword{internal}
