% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestR-package.R
\docType{package}
\name{nestR-package}
\alias{nestR}
\alias{nestR-package}
\title{nestR: Estimation of Bird Nesting from Tracking Data}
\description{
A package to locate nests and estimate avian reproductive outcome from GPS-tracking data. It includes functions to identify nest sites along bird movement trajectories based on patterns of location revisitation, dynamically explore results, estimate the outcome of identified nesting attempts, and plot and summarize outputs.
}
\author{
\strong{Maintainer}: Simona Picardi \email{simona.picardi@usu.edu}

Authors:
\itemize{
  \item Brian Smith \email{bsmity13@gmail.com}
  \item Matthew Boone \email{m.boone@ufl.edu}
  \item Mathieu Basille \email{basille@ufl.edu}
}

}
\keyword{internal}
