% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{nestr}
\alias{nestr}
\alias{nestr-package}
\title{Package Title}
\description{
Package description
}
\details{
The core functions/methods of this package:
 \itemize{
   \item{\code{\link[nestr]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
   \item{\code{\link[nestr]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/nestr}
}

